Version 0.6
Changed AddNewWindow to be in a new thread so I can set Apartmentstate
to STA
Changed structure of windows so that each window(with the exception of
overflow windows) is its own queue, meaning two windows will write
simultaneously instead of in order.
Added functionality to prevent stack overflows. It now creates a new
window, which is still referenced by the same key and writes to the new
window instead. Hopefully this will solve issues writing huge amounts of
text. The limit is currently set to 10k characters.
Fixed a bug where if the write process finished too quickly, the wait
line would throw an error.
Fixed a bug where IsNullOrEmpty would throw an error due to the fact
that items were still being added to the queue